requirements.txt
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
